[FEAT] Result 및 error 관련 엔티티 생성
