+++ src/main/java/capstone/ses/dto/system/ResultCode.java
@@ -1,4 +1,4 @@
-package capstone.ses.dto;
+package capstone.ses.dto.system;
 
 public enum ResultCode {
     SUCCESS, FAIL
